LayoutTests/http/tests/storageAccess/request-storage-access-same-origin-sandboxed-iframe-without-allow-token-expected.txt

CONSOLE MESSAGE: line 68: PASS document.hasStorageAccess was denied.
Tests that same-origin iframe storage access is denied if the iframe is sandboxed but doesn't have the allow token.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
